Add troubleshooting tip for `System.TypeLoadException` (#2969)
* Add troubleshooting tip for `System.TypeLoadException`

When running AspNet Core 1.1 on top of .NET 4.6.2 and trying to use the Azure Key Vault configuration extensions, I was encountering an exception that was being thrown for `System.TypeLoadException`. I discovered this originated from the use of the Azure Key Vault API and how it interacted with an older (built-in) version of `System.Net.HTTP`. I ended up installing the updated NuGet package for `System.Net.Http` and that fixed the issue. Hopefully this proposed tip will help someone in the future. Or at least until the packages get more aligned.

* Update key-vault-configuration.md