C19396: Missing hyperlink due to incorrect redenring (#6878)
Hello, @Rick-Anderson,

Localization team has reported source content issue that causes localized version to have broken format compared to en-us version.  
From the context of this source string, we can assume that writer did not intend this example URL to be rendered as hyperlink, and that text "App Namespace" within brackets {} is expected to be translated.

We found out that OPS is using a Markdig flavor called "Markdig (DocFX)" which does render URL surrounded by space to be rendered as hyperlink. 

To avoid this, suggested fix to source string is add an escape char (\) to part of the URL.
Please review and merge the proposed file change to fix to target versions. If you make related fix in another PR  then share your PR number so we can confirm and close this PR.

Many thanks in advance.