Added note for working with virtual directories (#9128)
I spent a day or more researching this so that I could use Swagger-UI in my ASP.NET Core API which is used with a reverse-proxy and virtual directory.  In the end found Swagger-UI requires relative URLs to start with a period.  Related to https://github.com/swagger-api/swagger-ui/issues/4107

Hoping this documentation note will help others.

When creating a new PR, please do the following and delete this template text:

* Reference the issue number if there is one:

  Fixes #Issue_Number

  > The "Fixes #nnn" syntax in the PR description causes
  > GitHub to automatically close the issue when this PR is merged.