SEO April: include ASP.NET Core 3.0 in title
SEO April Effort: This topic is for ASP.NET Core 3.0 and a top keyword being used in search included "3".  It is not immediately clear the topic is for ASP.NET Core 3.0 in search or even once you look at the topic.  Changing to see if the bounce rate improves.  Following up on bounce ratings next month for this as part of SEO hackadoc.