Update Create.cshtml.cs (#22963)
This change is necessary in case await _context. SaveChangesAsync (); will throw an exception and you will need to return the page back with the previously entered data, so as not to start typing again. If you leave it as it is, then when you try to save the changes, it gives an error that "An unhandled exception occurred while processing the request.
ArgumentNullException: Value cannot be null. (Parameter 'source')"